new plumlee script
this script is designed to start using a new method in surmise designed for stochastic simulations.  Right now, the script prints a ratio between the variance of function and the standard errors.  It shows something odd: more variation is present in the standard errors than the actual simulations.  Something must be wrong.